# NodeSetEditor.Server/OpenApi/v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "OPC UA NodeSet Editor API",
    "description": "REST API for the OPC UA NodeSet Editor.\n\nMost endpoints act on a workspace identified by the `OpcUa-Server` header, and NodeIds appearing in a URL path are base64url-encoded rather than percent-encoded.\n\nThis document is generated from the controllers at build time and checked into the repository; it is not produced at runtime.",
    "contact": {
      "name": "OPC Foundation",
      "url": "https://opcfoundation.org"
    },
    "version": "v1"
  },
  "paths": {
    "/api/opcua/v1/conformance-units": {
      "get": {
        "tags": [
          "ConformanceUnits"
        ],
        "parameters": [
          {
            "name": "modelId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "OpcUa-Server",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConformanceUnitInfo"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConformanceUnitInfo"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConformanceUnitInfo"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "AzureAd": [ ]
          },
          {
            "EmailCookie": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConformanceUnitInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "nodeCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "AzureAd": {
        "type": "http",
        "description": "Azure AD access token — the \"sign in with Microsoft\" path.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "EmailCookie": {
        "type": "apiKey",
        "description": "Session cookie issued by the passwordless email-code sign-in (`POST /api/auth/verify-code`). Sent automatically by the browser.",
        "name": "opc-email-auth",
        "in": "cookie"
      }
    }
  },
...
}